docs(project_rules): 更新项目编译运行和安装说明
更新项目规则文档中的编译运行和安装步骤，添加安装Debug包到真机的命令

fix(ui): 移除ChatActivity中冗余的cell.update调用
修复ChatActivity中选中消息时不必要的UI更新操作

refactor(ui): 优化ChatMessageCell中锁图标绘制
将锁图标绘制的参数类型转换为int以避免潜在精度问题

+++ TMessagesProj/src/main/java/org/telegram/ui/Cells/ChatMessageCell.java
@@ -22555,7 +22555,7 @@ public class ChatMessageCell extends BaseCell implements SeekBar.SeekBarDelegate
                     lockX = timeTitleTimeX + additionalX - iconSize - dp(2);
                     lockY = layoutHeight - dp(pinnedBottom || pinnedTop ? 7.5f : 6.5f) + timeYOffset - iconSize / 2;
                 }
-                setDrawableBounds(lockIcon, (int) lockX, (int) lockY, iconSize, iconSize);
+                setDrawableBounds(lockIcon, (int) lockX, (int) lockY, (int) iconSize, (int) iconSize);
                 lockIcon.setAlpha((int) (255 * alpha));
                 lockIcon.draw(canvas);
                 lockIcon.setAlpha(255);
